feat(theoros_remove_db): utils -> pragma_utils
