+++ rust/Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["pragma-cli", "theoros"]
+members = ["pragma-cli", "theoros", "pragma-utils"]
 
 [workspace.package]
 version = "0.1.0"
