+++ rust/pragma-cli/Cargo.toml
@@ -35,4 +35,4 @@ toml = { workspace = true }
 tracing = { workspace = true }
 url = { workspace = true }
 
-utils = { path = "../utils" }
+pragma-utils = { path = "../pragma-utils" }
