+++ rust/pragma-utils/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod conversions;
+pub mod services;
 pub mod tracing;
