+++ rust/theoros/Cargo.toml
@@ -33,4 +33,4 @@ utoipa = { workspace = true, features = ["axum_extras", "chrono", "uuid"] }
 utoipa-swagger-ui = { workspace = true, features = ["axum"] }
 utoipauto = { workspace = true }
 
-utils = { path = "../utils" }
+pragma-utils = { path = "../pragma-utils" }
