+++ rust/theoros/src/main.rs
@@ -10,11 +10,15 @@ use std::sync::Arc;
 use anyhow::Result;
 use prometheus::Registry;
 use tracing::Level;
-use utils::tracing::init_tracing;
+
+use pragma_utils::{
+    services::{Service, ServiceGroup},
+    tracing::init_tracing,
+};
 
 use crate::{
     config::config,
-    services::{ApiService, IndexerService, MetricsService, Service, ServiceGroup},
+    services::{ApiService, IndexerService, MetricsService},
     types::EventStorage,
 };
 
@@ -25,10 +29,9 @@ const EVENTS_MEM_SIZE: usize = 10;
 const METRICS_PORT: u16 = 8080;
 
 #[derive(Clone)]
-#[allow(unused)]
 pub struct AppState {
-    event_storage: Arc<EventStorage>,
-    metrics_registry: Registry,
+    pub event_storage: Arc<EventStorage>,
+    pub metrics_registry: Registry,
 }
 
 #[tokio::main]
@@ -45,9 +48,11 @@ async fn main() -> Result<()> {
     // TODO: state should contains the rpc_client to interact with a Madara node
     let state = AppState { event_storage: Arc::new(event_storage), metrics_registry: metrics_service.registry() };
 
-    // TODO: key in config
+    // TODO: Should be Pragma X DNA url - see with Apibara team + should be in config
+    let apibara_uri = "https://mainnet.starknet.a5a.ch";
+    // TODO: key in config / .env (...)
     let apibara_api_key = std::env::var("APIBARA_API_KEY")?;
-    let indexer_service = IndexerService::new(state.clone(), apibara_api_key);
+    let indexer_service = IndexerService::new(state.clone(), apibara_uri, apibara_api_key)?;
     let api_service = ApiService::new(state.clone(), config.server_host(), config.server_port());
 
     let theoros = ServiceGroup::default().with(metrics_service).with(indexer_service).with(api_service);
